+++ lib/sidekiq/rails.rb
@@ -52,11 +52,19 @@ module Sidekiq
       Sidekiq.hook_rails!
     end
 
-    # This hook happens after all initialziers are run, just before returning
-    # from config/environment.rb back to sidekiq/cli.rb.
     config.after_initialize do
-      if ::Rails::VERSION::MAJOR >= 5
-        Sidekiq.options[:reloader] = Sidekiq::Rails::Reloader.new
+      # This hook happens after all initializers are run, just before returning
+      # from config/environment.rb back to sidekiq/cli.rb.
+      # We have to add the reloader after initialize to see if cache_classes has
+      # been turned on.
+      #
+      # None of this matters on the client-side, only within the Sidekiq process itself.
+      #
+      Sidekiq.configure_server do |_|
+        if ::Rails::VERSION::MAJOR >= 5
+          Sidekiq.options[:reloader] = Sidekiq::Rails::Reloader.new
+          Psych::Visitors::ToRuby.prepend(Sidekiq::Rails::PsychAutoload)
+        end
       end
     end
 
@@ -73,6 +81,14 @@ module Sidekiq
 
       def inspect
         "#<Sidekiq::Rails::Reloader @app=#{@app.class.name}>"
+      end
+    end
+
+    module PsychAutoload
+      def resolve_class(klass_name)
+        klass_name && klass_name.constantize
+      rescue NameError
+        super
       end
     end
   end if defined?(::Rails)